+++ ExoPlayerAdapter/src/exo-analytics-listener-just-2_17/java/com/mux/stats/sdk/muxstats/internal/ExoAnalyticsListener.kt
@@ -83,10 +83,10 @@ private class ExoAnalyticsListener(player: ExoPlayer, val collector: MuxStateCol
 
   override fun onTimelineChanged(eventTime: AnalyticsListener.EventTime, reason: Int) {
     val player = player // strong reference during the listener call
-    if (player != null) {
-      Timeline.Window().apply {
-        eventTime.timeline.getWindow(0, this)
-        collector.sourceDurationMs = durationMs
+    if (player != null && eventTime.timeline.windowCount > 0) {
+        Timeline.Window().apply {
+          eventTime.timeline.getWindow(0, this)
+          collector.sourceDurationMs = durationMs
       }
     }
   }