+++ ExoPlayerAdapter/src/main/java/com/mux/stats/sdk/muxstats/internal/Util.kt
@@ -169,7 +169,7 @@ internal fun MuxStateCollector.handleExoPlaybackException(e: ExoPlaybackExceptio
 /////////////////////////////////////////////////////////////
 /// ExoPlayer Helper Functions //////////////////////////////
 
-fun ExoPlayer.MuxMediaHasVideoTrack(): Boolean {
+internal fun ExoPlayer.MuxMediaHasVideoTrack(): Boolean {
   val trackGroups:TrackGroupArray = getCurrentTrackGroups();
   var playItemHaveVideoTrack = false;
   if (trackGroups.length > 0) {