+++ MuxExoPlayer/src/main/java/com/mux/stats/sdk/muxstats/MuxStatsExoPlayer.kt
@@ -270,33 +270,45 @@ class MuxStatsExoPlayer @JvmOverloads constructor(
   inline fun <reified T : View> getExoPlayerView(): T? = playerView as? T
 
   /**
-   * Overwrite the OS version reported on the Mux Data dashboard for views recorded with this object
+   * Overwrite the device category
    */
+  @Deprecated(
+    message = "This method is not used by MuxCore any more",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
   fun overwriteOsVersion(osVersion: String) {
-    MuxDevice.muxStatsInstance?.overwrittenOsVersion = osVersion
   }
 
   /**
    * Overwrite the device name reported on the Mux Data dashboard for views recorded with this object
    */
+  @Deprecated(
+    message = "This method is not used by MuxCore any more",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
   fun overwriteDeviceName(deviceName: String) {
-    MuxDevice.muxStatsInstance?.overwrittenDeviceName = deviceName
   }
 
   /**
    * Overwrite the OS Family name (Such as 'Android')  reported on the Mux Data dashboard for views
    * recorded with this object
    */
+  @Deprecated(
+    message = "This method is not used by MuxCore any more",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
   fun overwriteOsFamily(osFamily: String) {
-    MuxDevice.muxStatsInstance?.overwrittenOsFamilyName = osFamily
   }
 
   /**
    * Overwrite the OS Family name (Such as 'Android')  reported on the Mux Data dashboard for views
    * recorded with this object
    */
+  @Deprecated(
+    message = "This method is not used by MuxCore any more",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
   fun overwriteManufacturer(manufacturer: String) {
-    MuxDevice.muxStatsInstance?.overwrittenManufacturer = manufacturer
   }
 
   /**
@@ -305,6 +317,7 @@ class MuxStatsExoPlayer @JvmOverloads constructor(
    */
   fun updateCustomerData(customerData: CustomerData) {
     muxStats.customerData = customerData
+    muxStats.updateHostDeviceData()
   }
 
   /**
@@ -499,11 +512,6 @@ private class MuxDevice(ctx: Context) : IDevice {
   private var appName = ""
   private var appVersion = ""
 
-  var overwrittenDeviceName: String? = null
-  var overwrittenOsFamilyName: String? = null
-  var overwrittenOsVersion: String? = null
-  var overwrittenManufacturer: String? = null
-
   override fun getHardwareArchitecture(): String {
     return Build.HARDWARE
   }
@@ -512,25 +520,57 @@ private class MuxDevice(ctx: Context) : IDevice {
     return "Android"
   }
 
-  override fun getMuxOSFamily(): String? = overwrittenOsFamilyName
+  @Deprecated(
+    message = "Mux core does not use this value anymore.",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
+  override fun getMuxOSFamily(): String? = ""
 
   override fun getOSVersion(): String {
     return Build.VERSION.RELEASE + " (" + Build.VERSION.SDK_INT + ")"
   }
 
-  override fun getMuxOSVersion(): String? = overwrittenOsVersion
+  @Deprecated(
+    message = "Mux core does not use this value anymore.",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
+  override fun getMuxOSVersion(): String? = ""
+
+  @Deprecated(
+    message = "Mux core does not use this value anymore.",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
+  override fun getMuxDeviceName(): String = ""
+
+  @Deprecated(
+    message = "Mux core does not use this value anymore.",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
+  override fun getMuxDeviceCategory(): String = ""
+
+  override fun getDeviceName(): String = ""
+
+  override fun getDeviceCategory(): String = ""
 
   override fun getManufacturer(): String {
     return Build.MANUFACTURER
   }
 
-  override fun getMuxManufacturer(): String? = overwrittenManufacturer
+  @Deprecated(
+    message = "Mux core does not use this value anymore.",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
+  override fun getMuxManufacturer(): String? = ""
 
   override fun getModelName(): String {
     return Build.MODEL
   }
 
-  override fun getMuxModelName(): String? = overwrittenDeviceName
+  @Deprecated(
+    message = "Mux core does not use this value anymore.",
+    replaceWith = ReplaceWith("CustomerViewerData")
+  )
+  override fun getMuxModelName(): String? = ""
 
   override fun getPlayerVersion(): String {
     return ExoPlayerLibraryInfo.VERSION
@@ -607,6 +647,17 @@ private class MuxDevice(ctx: Context) : IDevice {
     return SystemClock.elapsedRealtime()
   }
 
+  override fun outputLog(logPriority: LogPriority, tag: String, msg: String, t: Throwable?) {
+    when (logPriority) {
+      LogPriority.ERROR -> Log.e(tag, msg, t)
+      LogPriority.WARN -> Log.w(tag, msg, t)
+      LogPriority.INFO -> Log.i(tag, msg, t)
+      LogPriority.DEBUG -> Log.d(tag, msg, t)
+      LogPriority.VERBOSE -> Log.v(tag, msg, t)
+      else -> Log.v(tag, msg, t)
+    }
+  }
+
   override fun outputLog(logPriority: LogPriority, tag: String, msg: String) {
     when (logPriority) {
       LogPriority.ERROR -> Log.e(tag, msg)
@@ -638,7 +689,7 @@ private class MuxDevice(ctx: Context) : IDevice {
     const val CONNECTION_TYPE_OTHER = "other"
     const val MUX_DEVICE_ID = "MUX_DEVICE_ID"
 
-    val muxStatsInstance = MuxStats.getHostDevice() as? MuxDevice
+    val muxStatsInstance get() = MuxStats.getHostDevice() as? MuxDevice
   }
 
   init {